+++ stroom-data/stroom-data-store-impl-fs/src/main/java/stroom/data/store/impl/fs/FsOrphanFileFinderExecutor.java
@@ -20,14 +20,9 @@ package stroom.data.store.impl.fs;
 import stroom.data.store.impl.fs.shared.FindFsVolumeCriteria;
 import stroom.data.store.impl.fs.shared.FsVolume;
 import stroom.data.store.impl.fs.shared.FsVolume.VolumeUseStatus;
-import stroom.meta.shared.Meta;
-import stroom.task.api.ExecutorProvider;
 import stroom.task.api.TaskContext;
 import stroom.task.api.TaskContextFactory;
-import stroom.task.api.ThreadPoolImpl;
-import stroom.task.shared.ThreadPool;
 import stroom.util.io.FileUtil;
-import stroom.util.io.StreamUtil;
 import stroom.util.logging.LambdaLogger;
 import stroom.util.logging.LambdaLoggerFactory;
 import stroom.util.logging.LogExecutionTime;
@@ -35,15 +30,11 @@ import stroom.util.logging.LogExecutionTime;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.io.PrintWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Duration;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
 import java.util.function.Consumer;
 import javax.inject.Inject;
 import javax.inject.Provider;
@@ -60,21 +51,21 @@ class FsOrphanFileFinderExecutor {
     private final FsVolumeService volumeService;
     private final Duration oldAge;
     private final Provider<FsOrphanFileFinder> orphanFileFinderProvider;
-    private final ExecutorProvider executorProvider;
+//    private final ExecutorProvider executorProvider;
     private final TaskContextFactory taskContextFactory;
-    private final DataStoreServiceConfig config;
+//    private final DataStoreServiceConfig config;
 
     @Inject
     FsOrphanFileFinderExecutor(final FsVolumeService volumeService,
                                final Provider<FsOrphanFileFinder> orphanFileFinderProvider,
-                               final ExecutorProvider executorProvider,
+//                               final ExecutorProvider executorProvider,
                                final TaskContextFactory taskContextFactory,
                                final DataStoreServiceConfig config) {
         this.volumeService = volumeService;
         this.orphanFileFinderProvider = orphanFileFinderProvider;
-        this.executorProvider = executorProvider;
+//        this.executorProvider = executorProvider;
         this.taskContextFactory = taskContextFactory;
-        this.config = config;
+//        this.config = config;
 
         Duration age;
         age = config.getFileSystemCleanOldAge().getDuration();
@@ -104,28 +95,30 @@ class FsOrphanFileFinderExecutor {
 
         final List<FsVolume> volumeList = volumeService.find(FindFsVolumeCriteria.matchAll()).getValues();
         if (volumeList != null && volumeList.size() > 0) {
-            // Add to the task steps remaining.
-            final ThreadPool threadPool = new ThreadPoolImpl(TASK_NAME + "#",
-                    1,
-                    1,
-                    config.getFileSystemCleanBatchSize(),
-                    Integer.MAX_VALUE);
-            final Executor executor = executorProvider.get(threadPool);
-
-            final CompletableFuture<?>[] completableFutures = new CompletableFuture<?>[volumeList.size()];
+//            // Add to the task steps remaining.
+//            final ThreadPool threadPool = new ThreadPoolImpl(TASK_NAME + "#",
+//                    1,
+//                    1,
+//                    config.getFileSystemCleanBatchSize(),
+//                    Integer.MAX_VALUE);
+//            final Executor executor = executorProvider.get(threadPool);
+//
+//            final CompletableFuture<?>[] completableFutures = new CompletableFuture<?>[volumeList.size()];
             int i = 0;
             for (final FsVolume volume : volumeList) {
                 if (VolumeUseStatus.ACTIVE.equals(volume.getStatus())) {
-                    final Runnable runnable = taskContextFactory.childContext(parentContext,
-                            "Checking: " + volume.getPath(),
-                            taskContext ->
-                                    scanVolume(volume, orphanConsumer, oldestDirTime, taskContext));
-                    final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable,
-                            executor);
-                    completableFutures[i++] = completableFuture;
+                    scanVolume(volume, orphanConsumer, oldestDirTime, parentContext);
+
+//                    final Runnable runnable = taskContextFactory.childContext(parentContext,
+//                            "Checking: " + volume.getPath(),
+//                            taskContext ->
+//                                    scanVolume(volume, orphanConsumer, oldestDirTime, taskContext));
+//                    final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable,
+//                            executor);
+//                    completableFutures[i++] = completableFuture;
                 }
             }
-            CompletableFuture.allOf(completableFutures).join();
+//            CompletableFuture.allOf(completableFutures).join();
         }
 
         parentContext.info(() -> "start() - Completed orphan file finder in " + logExecutionTime);