+++ CHANGELOG.md
@@ -17,6 +17,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Change data viewing 'progress' bar to have a minimum width of 3px to make it more visible.
 
+* Issue **#2469** : Made `SafeXmlFilter` available for use in the application.
+
+* Issue **#2403** : Fix guice bind errors in stroom and proxy when the `path` config branch is empty or set to null.
+
+* Issue **#2474** : Added debug to help diagnose issue with orphan file finder.
+
 * Issue **#2462** : Made changes to stop DB connections being used within the context of other open connections.
 
 * Uplift LMDBJava to 0.8.2 to fix LMDBJava cursor comparator bug.